Store roles through server list query (#1489)
