+++ webui/src/pages/Cluster/components/ServerDetailsModal/components/IssuesTab/IssuesTab.tsx
@@ -1,12 +1,12 @@
 import React, { memo } from 'react';
 import { Text } from '@tarantool.io/ui-kit';
 
-import type { ServerListServerClusterIssue } from 'src/models';
+import type { ServerListClusterIssue } from 'src/models';
 
 import { styles } from './IssuesTab.styles';
 
 export interface IssuesTabProps {
-  issues: ServerListServerClusterIssue[];
+  issues: ServerListClusterIssue[];
 }
 
 const IssuesTab = ({ issues }: IssuesTabProps) => {
